requirements.txt
opencv-python
fastapi
uvicorn[standard]
influxdb-client
python-dotenv